+++ src/Pt-XmlRpc/Formatter.cpp
@@ -69,7 +69,7 @@ void Formatter::addBool(const std::string& name, bool value,
 }
 
 
-void Formatter::addInt(const std::string& name, long value, 
+void Formatter::addInt(const std::string& name, long long value, 
                        const std::string& id)
 {
 	convert(_value, value);
@@ -77,7 +77,7 @@ void Formatter::addInt(const std::string& name, long value,
 }
 
 
-void Formatter::addUInt(const std::string& name, unsigned long value, 
+void Formatter::addUInt(const std::string& name, unsigned long long value, 
                         const std::string& id)
 {
 	convert(_value, value);
